_layout.tsx holds each page with buttons, still need to change svg, however index holds the home page
