--- FrugalApp/app/(tabs)/_layout.tsx
@@ -1,16 +1,0 @@
-import { Tabs } from 'expo-router';
-import React from 'react';
-import { Platform, StyleSheet } from 'react-native';
-import { useColorScheme } from '@/hooks/useColorScheme';
-import { View } from 'react-native';
-
-export default function TabLayout() {
-  const colorScheme = useColorScheme();
-
-  return (
-      <Tabs screenOptions={{ headerShown: false, tabBarStyle: { display: 'none' } }}>
-        <Tabs.Screen name="home" options={{ title: "Home" }} />
-      </Tabs>
-  );
-}
-
