+++ FrugalApp/app/(tabs)/index.tsx
@@ -1,60 +1,169 @@
-import { Image, StyleSheet, Platform, Button } from "react-native";
-import { View, Text, TextInput } from "react-native";
-import { useState } from "react";
+import React, { useState } from 'react';
+import { View, Text, TextInput, Image, ScrollView, FlatList, TouchableOpacity, StyleSheet } from 'react-native';
+import { SafeAreaView } from 'react-native-safe-area-context';
+import { Ionicons } from '@expo/vector-icons';
+import { useColorScheme } from '@/hooks/useColorScheme';
+
+const categories = [
+  { id: 1, name: 'Vegetables', image: require('@/assets/images/veggies.svg') },
+  { id: 2, name: 'Fruits', image: require('@/assets/images/fruits.svg') },
+  { id: 3, name: 'Beverages', image: require('@/assets/images/beverages.svg') },
+  { id: 4, name: 'Grocery', image: require('@/assets/images/oil.svg') },
+  { id: 5, name: 'Household', image: require('@/assets/images/household.svg') },
+];
+
+const products = [
+  { id: 1, name: 'Fresh Peach', price: '$5.00 - $8.00', image: require('@/assets/images/banana.jpg') },
+  { id: 2, name: 'Organic Apple', price: '$3.00 - $5.00', image: require('@/assets/images/banana.jpg') },
+  { id: 3, name: 'Banana Bunch', price: '$2.00 - $4.00', image: require('@/assets/images/banana.jpg') },
+  { id: 4, name: 'Carrot Pack', price: '$1.50 - $3.00', image: require('@/assets/images/banana.jpg') },
+  { id: 5, name: 'Tomato Pack', price: '$2.50 - $4.50', image: require('@/assets/images/banana.jpg') },
+  { id: 6, name: 'Broccoli Bunch', price: '$3.00 - $5.00', image: require('@/assets/images/banana.jpg') },
+];
 
 export default function HomeScreen() {
-  const [text, setText] = useState("");
-
+  const [searchText, setSearchText] = useState('');
+  const theme = useColorScheme() ?? 'dark';
 
   return (
-    <View style={styles.container}>
-      <TextInput
-      style={styles.input}
-      placeholder="Write here"
-      value={text}
-      onChangeText={setText}
+    <SafeAreaView style={styles.container}>
+      {/* Search Bar */}
+      <View style={styles.searchBar}>
+        <Ionicons name="search" size={24} color="gray" style={styles.searchIcon} />
+        <TextInput
+          placeholder="Search keywords..."
+          style={styles.searchInput}
+          value={searchText}
+          onChangeText={setSearchText}
+        />
+      </View>
+
+      {/* Ready to Save Banner */}
+      <View style={styles.readyToSaveContainer}>
+        <Text style={styles.readyToSaveText}>Ready to Save?</Text>
+        <Image source={require('@/assets/images/banana.jpg')} style={styles.readyToSaveImage} />
+      </View>
+
+      {/* Categories Section */}
+      <View style={styles.bannerContainer}>
+        <Text style={styles.bannerText}>Categories</Text>
+      </View>
+
+      {/* Categories */}
+      <ScrollView horizontal showsHorizontalScrollIndicator={false} style={styles.categoriesContainer}>
+        {categories.map((cat) => (
+          <TouchableOpacity key={cat.id} style={styles.categoryItem}>
+            <Image source={cat.image} style={styles.categoryImage} />
+            <Text style={styles.categoryText}>{cat.name}</Text>
+          </TouchableOpacity>
+        ))}
+      </ScrollView>
+
+      {/* Featured Products */}
+      <Text style={styles.featuredProductsText}>Featured Products</Text>
+      <FlatList
+        data={products}
+        numColumns={2}
+        keyExtractor={(item) => item.id.toString()}
+        columnWrapperStyle={styles.productListColumn}
+        renderItem={({ item }) => (
+          <TouchableOpacity style={styles.productItem}>
+            <Image source={item.image} style={styles.productImage} />
+            <Text style={styles.productName}>{item.name}</Text>
+            <Text style={styles.productPrice}>{item.price}</Text>
+          </TouchableOpacity>
+        )}
       />
-
-      <View style={styles.button}>
-        <Button title="save" onPress={() => save(text)} />
-      </View>
-
-      <View style={styles.button}>
-        <Button title="load" />
-      </View>
-    </View>
+    </SafeAreaView>
   );
 }
 
-function save(text: string): void {
-  console.log("Save: ", text);
-}
-
 const styles = StyleSheet.create({
-  textPlain: {
-    color: "white",
-    fontSize: 18,
-    marginTop: 20,
-  },
-
-  button: {
-    margin: 10,
-  },
-
   container: {
     flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "white",
+    backgroundColor: 'white',
+    padding: 16,
   },
-
-  input: {
-    width: "80%",
+  searchBar: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    backgroundColor: '#f0f0f0',
+    padding: 8,
+    borderRadius: 16,
+  },
+  searchIcon: {
+    marginRight: 8,
+  },
+  searchInput: {
+    flex: 1,
+    color: '#666',
+  },
+  readyToSaveContainer: {
+    alignItems: 'center',
+    marginTop: 16,
+  },
+  readyToSaveText: {
+    fontSize: 22,
+    fontWeight: 'bold',
+    marginBottom: 8,
+  },
+  readyToSaveImage: {
+    width: '100%',
+    height: 160,
+    borderRadius: 8,
+  },
+  bannerContainer: {
+    marginTop: 16,
+  },
+  bannerText: {
+    fontSize: 20,
+    fontWeight: 'bold',
+  },
+  categoriesContainer: {
+    marginTop: 16,
+  },
+  categoryItem: {
+    alignItems: 'center',
+    marginHorizontal: 12,
+  },
+  categoryText: {
+    color: '#666',
+    fontSize: 12,
+    marginTop: 4,
+  },
+  categoryImage: {
+    width: 50,
     height: 50,
-    borderColor: "black",
-    borderWidth: 1,
-    padding: 10,
-    color: "black",
-    borderRadius: 5,
+    borderRadius: 25,
+    marginBottom: 4,
+  },
+  featuredProductsText: {
+    fontSize: 18,
+    fontWeight: '600',
+    marginTop: 24,
+  },
+  productListColumn: {
+    justifyContent: 'space-between',
+  },
+  productItem: {
+    backgroundColor: '#f0f0f0',
+    padding: 12,
+    borderRadius: 8,
+    marginTop: 16,
+    width: '48%',
+  },
+  productImage: {
+    width: '100%',
+    height: 112,
+    borderRadius: 8,
+  },
+  productName: {
+    fontSize: 14,
+    fontWeight: '500',
+    marginTop: 8,
+  },
+  productPrice: {
+    color: '#999',
+    fontSize: 12,
   },
 });
